# fix soundness concern in h1 decoder