--- a/actix-http/Cargo.toml
+++ b/actix-http/Cargo.toml
@@ -103,3 +103,7 @@ harness = false
 [[bench]]
 name = "status-line"
 harness = false
+
+[[bench]]
+name = "uninit-headers"
+harness = false