@Shnatsel 2020-08-17
This RFC will allow getting performance back safely: https://github.com/rust-lang/rfcs/pull/2930